+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 3.1.30 Jan 13, 2020
+
+Added `--minGenomeLength` to `make-protein-database.py`.
+
+## 3.1.28 Dec 20, 2019
+
+Removed the unused `taxonomy` (`VARCHAR`) column from the genomes table of
+the protein database.
+
 ## 3.1.27 Dec 3, 2019
 
 Fixed silly bug in alignment filtering code that was somehow not tested.